Send valid track number even when Music is paused
Send valid track number even when Music is paused.
Otherwise when CK connects for the first time to
DUT with song as paused, DUT shares the track id
as 0xFFFFFFFF which forces CK not to ask for meta
data of the current song and CK UI remains blank.
CK gets valid track id only when song is changed
from DUT and then only it can fetch the meta data
and update its UI.

Change-Id: If0a2b69c723133cfc2d8e2e6944813c037c0c560